Moved validation rules targeting document dto from input validation to document validation
