+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Domain/Dtos/ChargeCommands/Validation/DocumentValidation/ValidationRules/RecipientIsMandatoryTypeValidationRule.cs
@@ -12,24 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using GreenEnergyHub.Charges.Domain.Dtos.ChargeCommands.Validation.InputValidation.ValidationRules;
 using GreenEnergyHub.Charges.Domain.Dtos.SharedDtos;
 using GreenEnergyHub.Charges.Domain.Dtos.Validation;
-using GreenEnergyHub.Charges.Domain.MarketParticipants;
 
-namespace GreenEnergyHub.Charges.Domain.Dtos.ChargeCommands.Validation.InputValidation.ValidationRules
+namespace GreenEnergyHub.Charges.Domain.Dtos.ChargeCommands.Validation.DocumentValidation.ValidationRules
 {
-    public class DocumentTypeMustBeRequestUpdateChargeInformationRule : IValidationRule
+    public class RecipientIsMandatoryTypeValidationRule : IValidationRule
     {
         private readonly DocumentDto _documentDto;
 
-        public DocumentTypeMustBeRequestUpdateChargeInformationRule(DocumentDto documentDto)
+        public RecipientIsMandatoryTypeValidationRule(DocumentDto documentDto)
         {
             _documentDto = documentDto;
         }
 
         public ValidationRuleIdentifier ValidationRuleIdentifier =>
-            ValidationRuleIdentifier.DocumentTypeMustBeRequestUpdateChargeInformation;
+            ValidationRuleIdentifier.RecipientIsMandatoryTypeValidation;
 
-        public bool IsValid => _documentDto.Type == DocumentType.RequestUpdateChargeInformation;
+        public bool IsValid => MarketParticipantMrIdValidator.IsValid(_documentDto.Recipient.Id);
     }
 }
