+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Domain/GreenEnergyHub.Charges.Domain.csproj
@@ -42,4 +42,8 @@ limitations under the License.
       <ProjectReference Include="..\GreenEnergyHub.Charges.Core\GreenEnergyHub.Charges.Core.csproj" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Dtos\ChargeCommands\Validation\OperationValidation" />
+    </ItemGroup>
+
 </Project>
